Fix Interrupt example (#69)
* Fix Interrupt example

Use #define's for all port references, add some explanatory comment, and ensure the scheduled packet send uses the DEVPORT port as intended.

* Interim commit of SS

This doesn't work, but does compile and attempt to work. The ISR runs every ~2us and while I can turn it off and on, and my ISR runs, I can't seem to configure it as a timer properly. Ideal solution would be to set it as a timer for 104us (for 9600bps) and let other stuff happen inbetween the bits.