Update radio.c
Deleted RadioLoRaSymbTime - it's not used anymore ( replaced by RadioSymbTime() function )

return of RadioSymbTime is 2^SpreadingFactor / Frequency in kHz

For SF 11 and 12 and Bandwidth 125 the return is > 16.38. Same for Bandwidth 250 and SF 12. So the extensive checking in both changed IFs is not necessary, because the condition will always be met and LowDataRateOptimize will be set.

Changed "16.38" to "16.3" for the case SF=8 and BW=15.63kHz (result is 16.379)